Initial cleanup for the profiler agent method cache.
